Move lookup out of image content controller
