+++ applications/app/services/repositories.scala
@@ -8,6 +8,9 @@ import com.gu.openplatform.contentapi.model.ItemResponse
 import org.joda.time.DateTime
 import org.scala_tools.time.Implicits._
 import contentapi.QueryDefaults
+import controllers.ImageContentPage
+import scala.concurrent.Future
+import play.api.mvc.SimpleResult
 
 case class IndexPage(page: MetaData, trails: Seq[Trail])
 
@@ -44,5 +47,22 @@ trait Index extends ConciergeRepository with QueryDefaults {
     val latest: Seq[Content] = response.results.map(Content(_)).filterNot(c => leadContent.map(_.id).exists(_ == c.id))
     val allTrails = (leadContent ++ editorsPicks ++ latest).distinctBy(_.id).take(20)
     tag map { IndexPage(_, allTrails) }
+  }
+}
+
+trait ImageQuery extends ConciergeRepository with QueryDefaults {
+
+  def image(edition: Edition, path: String): Future[Either[ImageContentPage, SimpleResult]]= {
+    log.info(s"Fetching image content: $path for edition ${edition.id}")
+    val response = ContentApi.item(path, edition)
+      .showExpired(true)
+      .showFields("all")
+      .response.map { response =>
+      val mainContent: Option[Content] = response.content.filter { c => c.isImageContent } map {Content(_)}
+      val storyPackage: List[Trail] = response.storyPackage map { Content(_) }
+      mainContent.map { content => Left(ImageContentPage(content,storyPackage)) }.getOrElse(Right(NotFound))
+    }
+
+    response recover suppressApiNotFound
   }
 }
